requirements.txt
langgraph
langchain-google-genai
langchain-openai
langchain-mcp-adapters
pymongo
fastmcp
python-dotenv